+++ src/common/lib/directus.ts
@@ -1,4 +1,4 @@
-import { createDirectus, rest } from '@directus/sdk'
+import { createDirectus, rest, staticToken } from '@directus/sdk'
 
 if (!process.env.DIRECTUS_CMS_URL) {
   throw new Error('DIRECTUS_CMS_URL environment variable is not defined')
@@ -11,5 +11,9 @@ const directus = createDirectus(directusCmsUrl).with(
     onRequest: (options) => ({ ...options }),
   }),
 )
+
+export function createPreviewDirectus(token: string) {
+  return createDirectus(directusCmsUrl).with(staticToken(token)).with(rest())
+}
 
 export default directus